+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## X.Y.Z - changes pending release 
+## X.Y.Z - changes pending release
+
+### PaymentSheet
+* [Added] Support for PayPay with PaymentIntents.
+
+### StripePayments
+* [Added] Support for PayPay bindings.
 
 ### CustomerSheet
 * [Added] Added support for `onBehalfOf` to CustomerSheet.IntentConfiguration. This parameter makes CustomerSheet use a connected account to determine the payment method that users see and whether CardBrandChoice is enabled. For more information, see the [SetupIntent docs](https://docs.stripe.com/api/setup_intents/object#setup_intent_object-on_behalf_of).